Add files via upload
Glass Classification Project

This project involves classifying samples of glass based on their chemical properties. The dataset used in this project contains information about the chemical compositions of different types of glass, measured as weight percent in corresponding oxides. The goal is to build a model that can accurately classify the type of glass based on its chemical properties.

The dataset consists of 214 observations with 10 features, including the refractive index (RI) and the weight percentages of various oxides such as Sodium (Na), Magnesium (Mg), Aluminum (Al), Silicon (Si), Potassium (K), Calcium (Ca), Barium (Ba), and Iron (Fe). The target variable is the type of glass, which falls into one of seven classes.

The project includes several steps, such as data loading and preprocessing, feature scaling, model training using machine learning algorithms like Random Forest, and evaluating the model's performance. The accuracy of the model on the training data, as well as the classification report, is calculated to assess the model's effectiveness.

This project serves as a practical example of applying machine learning techniques to solve a classification problem in the domain of material science. It can be used as a starting point for further exploration, optimization, or development of more advanced glass classification models.